Ensure that the PolyglotListener is invoked after any other subscriber or listener
We've already encountered the issue, where the polyglot doctrine listener
would be executed before another one, that might've been responsible for
telling the entity on how to lazily fetch the data in the translatable
property.
That way the later listeners/subscribers will compromise the
proxy object for translation.

To fix this, we set the priority of the "doctrine.event_listener" tag
of the webfactory.polyglot.doctrine_listener service to -100.

What we didn't know at the time, is that event listeners,
 meaning services tagged with "doctrine.event_listener" will always be
registered after event subscribers ("doctrine.event_subscriber").

Now we've changed the implementation of the PolyglotListener
to be tagged as doctrine event listener instead of subscriber.