Annotate classes to be `@final`
This is a soft, friendly way to announce we're going to make classes final in the next major version. It should be picked up at least by tools like Psalm or decent IDEs.

All those classes were never  meant to be extended or to be used as extension points, only it was not technically prohibited.