Run functional tests without the doctrine-orm-infrastructure dependency (#85)
I am unsure about the future of the [Doctrine-ORM-Test-infrastructure
package](https://github.com/webfactory/doctrine-orm-test-infrastructure/):
Recent deprecations in ORM 3 and especially DBAL 3 make it seem harder
and harder to come up with something that closely resembles a working
ORM setup "out of nothing".

So, this PR gets rid of that dependency. It builds upon Symfonys
functional testing infrastructure, and configures a copy of the default
entity manager in the DIC. This is good enough to have separate channels
for importing/loading data and still having a pristine entity manager
afterwards (altough, admittedly, a `clear()` call would probably have
the same effect?).

Also, setting up the schema in memory is not all too complicated,
leveraging the ORM's schema tool.